Remove redundant nomenclature
